MEPTS-24: Refactoring -Introducing GenericIndicators and removing unused methods
